refactor(a2a-t-sdk): unify MetadataContent in core model
