+++ a2a-t-client/src/main/java/net/openan/a2at/sdk/client/prompt/orchestration/ClientPromptGenerationOrchestrator.java
@@ -1,7 +1,7 @@
 package net.openan.a2at.sdk.client.prompt.orchestration;
 
 import java.util.Map;
-import net.openan.a2at.sdk.client.model.MetadataContent;
+import net.openan.a2at.sdk.core.model.MetadataContent;
 import net.openan.a2at.sdk.client.model.PromptGenerationResult;
 
 /**
